KVM: add a memory-slot reference-count mechanism
Background:

Guest memory-mappings are maintained in translation tables called
"memory slots" in KVM.  Each slot may pertain to a specific type
of contiguous memory, such as RAM, BIOS-ROM, VGA buffers, etc.

Generally speaking, these slots can be "hotplugged" at any time,
thus dynamically shaping the memory map of the guest.  Certain types
of memory (such as RAM) will technically never change over the lifetime
of the guest even though the basic facility permits this.

Motivation:

We want to build high-performance IO subsystems to interact with guests.
In order to do this, we need to acquire long term references to guest
RAM physical addresses and map them into the host kernel.  Since
we know RAM addresses that are in use-by the guest should not be changing,
we can more or less do this today by simply translating a gpa-to-hva
and peforming a vmap.  However, there is no mechanism outside of holding
the slots_lock to protect access to the mappings.

Therefore, we introduce a reference counting mechanism simply for
informational purposes.  This will allow us to detect if a slot is
still in-use by a guest, and will print a warning on the host.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>