kvm: add io services to xinterface
We want to add a more efficient way to get PIO signals out of the guest,
so we add an "xioevent" interface.  This allows a client to register
for notifications when a specific MMIO/PIO address is touched by
the guest.  This is a higher-performance alternative to ioeventfd,
which is mired in eventfd overhead.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>