vbus: implement "fastcall" hypercalls in the pci-bridge
This breaks the existing bridge ABI to restructure the register layout
and to add a new feature called "fastcall".  This lets the guest
hypercall directly to the host kernel without any locks, and without
multiple-mmio hops to userspace.  Generally, most hypercalls are not
fastpath.  However, certain applications of DEVCALL will become fastpath
in the future (such as real-time support), so lets just fix this now
before the ABI is solid.

For simplicity, a SLOWCALL method is provided for guests which do not
want the extra complexity of setting up a fastcall-channel.  This allows
a hypercall to execute, albeit through the traditional synchronous,
multiple-mmio via userspace method.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>