KVM: convert xinterface to use use_mm instead of get_user_pages.
use_mm+copy_to/from_user is an optimized path that is much faster than
get_user_pages for cases where the page is already in memory.  Since
devices typically converse with memory recently touched, the probability
that a given page is already resident is fairly high.  Therefore, we
can get a performance boost if we happen to be calling from a context
that is conducive to using the copy_to/from_user variants (via current->mm
or use_mm).  As it so happens, most contexts will be in one or the
other category anyway, so this results in a large (~25%) throughput
improvement overall.

Numbers on 10GE 802.x ethernet jumped from 4560Gb/s to 5708Gb/s with this
patch.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>