venettap: add zerocopy support
This adds the ability to directly map a guests pages into a paged-skb
instead of copying the data, saving the overhead and cache-thrashing
normally associated with copying data into a host specific skb.

This feature requires "TXC" (transmit-complete events) support in the
guest to be enabled before zero-copy can be enabled.  This patch adds
a new attribute to the device-model called "zcthresh" which sets the
threshold between normal copy and zero-copy operation.  Setting this
threshold to '0' disables zero-copy.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>