vbus: add administrative ioctl interface to compliment configfs
This creates a new /dev/vbus-admin inode for controlling some
rudimentary elements of a virtual-bus configuration.  The idea
is that this inode can be assigned to a particular gid, and that
gid would then be authorized to create/manage vbus containers.

This buys us two things over the existing configfs interface:

1) Non-root users may now be granted rights to create vbus containers

2) The containers have a lifetime associated with the creator (i.e.
   auto-cleanup)

Signed-off-by: Gregory Haskins <ghaskins@novell.com>